separated portWatcher
